ogg: rfc5334: Detect audio/ogg and video/ogg
application/ogg used[1] to be the MIME type recommended for any Ogg
encapsulated file. This is obsoleted by the new RFC 5334[2].

audio/ogg is to be used for for all Ogg files that contain
predominantly audio, such as Ogg Vorbis files (.ogg or .oga)

video/ogg is to be used for all Ogg files that contain
predominantly video, such as Ogg Theora or Ogg Dirac files.
The file extension recommended to be used is .ogv

Reference:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Configuring_servers_for_Ogg_media#Serve_media_with_the_correct_MIME_type

[1]: https://www.ietf.org/rfc/rfc3534.txt
[2]: https://www.ietf.org/rfc/rfc5334.txt

Signed-off-by: Prashanth Pai <prashanth.pai@plivo.com>