Add go version to go.mod
On go1.13, every time `go build` or `go test` is run, the `go.mod`
file is automatically populated with a Go version. It will be set
to the current release (go1.13) if no existing version is present.

The version of Go in go.mod shall be the minimum version of Go that
the module supports.

Pardon my OCD by it is little unsettling to have to explicitly not
check in the go.mod file in the commit every time.

Signed-off-by: Prashanth Pai <prashanth.pai@plivo.com>