Refactor mso xml detection for trailing signatures
Zip archives do not care about the order of the files inside them. This
makes the signature files of Ms office formats to be in the trailing
part of the archive and not get picked up by the detection alg. This
commit improves detection by giving up search for general MS office
signature search in favor of searching for more precise entries.
Related issue: #81