Add readLimit param in matching function definition
Some detections functions need to know whether the input they are
processing is the whole file or just the header. Now they can find it
out by comparing len(in) with readLimit. readLimit is passed as
parameter and not read from the package variable because SetLimit can
concurrently alter it.