add application/xml as alias; close #227 (#581)
According to RFC7303, text/xml is an alias to application/xml. But
considering we we're using text/xml as the main mime type until now,
changing to main=application/xml alias=text/xml would cause trouble to
users. So for now, we're keeping as: main=text/xml alias=application/xml