tests: revert effects of Extend
previously: tests would alter the structure of the MIME tree and leave
it like that. The consequence was that, depending on tests order,
benchmarks, or the tests that generates supported_mimes.md would work
with a dirty MIME tree.

after change: MIME tree is left clean

It is still not ideal. Tests in go are supposed to run concurrently, but that's
troublesome when they all alter the same global MIME tree.