zip: stop searching when PK\x03\x04 cannot be found (#696)
This bug was discovered when detection started to take a long time for a
bigger zip file. Because zipIterator returned many bogus entries and KMZ
detection iterated over all of them, the detection time for zip files
increased noticeably.
The zip file that showed increased detection time is
https://github.com/gabriel-vasile/mimetype_tests/blob/0a0e22bb4af7dbd879a934a17ccc7f9ec89c633e/testfiles/format-corpus-master/droid-jhove-datetest-files/audio_historicaldates.zip
    
Additionally, this PR limits KMZ detection to first 100 entries in the
zip, same as all MS Office detections.