shell: compact whitespace when matching shebangs (#714)
* scan: make Search and Match return how many bytes they consumed

After Match matches a pattern, it's useful to know how many bytes from
the input it took.

* shell: compact white space when matching shebang

Before this commit, detection would fail for shebangs with insignificant
white space. Ex:
   #!/usr/bin/env     python
This commit makes it so whitespace is compacted when matching shebangs.

* charset: replace if with switch to please linter

* scan: add flag for matching full words

It makes Match return false when the matched pattern
is not followed by a space character.

* shebang: match full words for bash and incomplete for python

This commit uses the FullWord flag to make detection  behave more like file.
For bash shebangs:
 #!/usr/bin/env bash is matched successfully
 #!/usr/bin/env bash1.0.0 is failing
For python shebangs:
 #!/usr/bin/env python is matched successfully
 #!/usr/bin/env python3.10 is also matched successfully

The reason why shebangs are matched differently between bash and python
is because python shebangs can sometimes include the python version, as in:
 #!/usr/bin/env python3.10
while for bash it's not common to.

* ci: add missing fuzz functions

There must be a way to select and run all fuzzing functions without
listing them individually, but I couldn't find it.
Additionally, add a ref:master in benchmark workflow. I accidentally
removed it when I upgraded action versions. It's needed in order to
compare master benchmark with PR branch benchmark.

* benchmark: reduce runs from 10 to 7

There is not big gap in runtime between runs, so might as well make the
CI take less time.