aaf: remove individual node for aaf
When it was introduced in #137, it was because the matching function for
.doc was a catch-all return true: all ole files that we could not
identify as xls, ppt, etc. were considered doc.
Today, that doc is not a catch-all anymore, so .aaf files are detected
as x-ole-storage, even when specialised detector for aaf does not exist.
The problem with specialised detector: aaf is not a documented format
with an assigned Media type. We we're re-using application/octet-storage
and that is confusing.