msoxml: match files and directories for first zip entry
Previously, bytes.Equal was used for comparing first entry from the zip.
That worked most of the times (because [Content_Types].xml, i.e. a file,
not a dir) is the first entry in the zip. But for docProps, customXml,
etc. the check was failing because those are directories. The failing
check was, for example, bytes.Equal("docProps/app.xml", "docProps").