json: abort fast when expected first token does not match
Previously, a check for object or array was done. But this is not ideal
for a case where a huge array is provided as input, but we're only interested
in JSON objects.
This commit makes jsonHelper abort as soon as the first token is not what
we expect. That saves us from parsing whole JSON inputs.