zip+json: performance improvements (#732)
* json: abort fast when expected first token does not match

Previously, a check for object or array was done. But this is not ideal
for a case where a huge array is provided as input, but we're only interested
in JSON objects.
This commit makes jsonHelper abort as soon as the first token is not what
we expect. That saves us from parsing whole JSON inputs.

* zip: fix error logic when matching files in dir entries

Previously, for zip dir check and in the negative case, both bytes.HasPrefix
and bytes.Equal were called. bytes.Equal was redundant: it was failing
always when bytes.HasPrefix failed.
Additionally, jar detection is improved to match just a directory (previous
check for file was redundant.)

* fuzz: use latest stable go