all: check for integer overflows
Sometimes we need to look for a value inside the input and the offset of
that value comes from the input too. In those cases an integer overflow
chance exists. When overflow happens an index out of range error occurs,
as shows in #731.

This commit abort checks when an overflow is detected.